requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
DateTime==4.3
discord.py==1.3.3
distlib==0.3.0
dnspython==2.0.0
filelock==3.0.12
identify==1.4.15
idna==2.9
joblib==0.14.1
lxml==4.5.0
multidict==4.7.5
nodeenv==1.3.5
numpy==1.18.4
packaging==20.3
pluggy==0.13.1
pre-commit==2.3.0
profanity-check==1.0.3
py==1.8.1
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.3.0
pyparsing==2.4.7
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
soupsieve==2.0
toml==0.10.0
tox==3.15.0
translate==3.5.0
urllib3==1.25.9
virtualenv==20.0.20
websockets==8.1
yarl==1.4.2
youtube-dl==2020.3.24
zope.interface==5.1.0
